Syntatic and Lexical Parsers base

+++ ArithmeticParser/Utils.cpp
@@ -4,7 +4,7 @@
 
 void RemoveCharactersFromString(std::string& s, const char* characters)
 {
-	for (unsigned int i = 0; i < strlen(characters); ++i)
+	for (size_t i = 0; i < strlen(characters); ++i)
 	{
 		s.erase(std::remove(s.begin(), s.end(), characters[i]), s.end());
 	}
